+++ client/src/types/hooks/forms.ts
@@ -1,1 +1,7 @@
 export type Fields = Record<string, string>
+
+
+export type LoginFormFields = {
+    username : string,
+    password : string
+}
